Refactor the occurrences schema by splitting it to use new services file
PT-1651.
Added a new `enrol_to_occurrences` service function
to handle the enrolling to a number of occurrences.
fix
