+++ occurrences/tests/snapshots/snap_test_api.py
@@ -4,6 +4,7 @@ from __future__ import unicode_literals
 
 from snapshottest import Snapshot
 
+
 snapshots = Snapshot()
 
 snapshots["test_add_occurrence_to_published_event 1"] = {
